[optimisations] added distance based density to grass

+++ source/runtime/Geometry/GeometryGeneration.h
@@ -591,7 +591,7 @@ namespace spartan::geometry_generation
             float sa             = std::sin(angle);
             float cos_t          = std::cos(this_tilt);
             float sin_t          = std::sin(this_tilt);
-            uint32_t petal_start = vertices->size();
+            uint32_t petal_start = static_cast<uint32_t>(vertices->size());
             for (uint32_t i = 0; i <= petal_segment_count; ++i)
             {
                 float t = static_cast<float>(i) / petal_segment_count;
